requirements.txt
Django>=4.2.0
reportlab==4.1.0
google-generativeai==0.3.2
python-dotenv==1.0.1
Pillow==10.2.0  # Required for reportlab 